requirements.txt
discord_webhook==1.4.1
python-dotenv==1.1.0
Requests==2.32.3
